Move summary plot tests to new file
